tegra: hdmi: remove unnecessary call to trusty
The HDCP 1.4 can't be recovered when hot-plug off then on because
the trusted session is abnormally closed (in the middle of hdcp is working)
by unnecessary nvhdcp_te_open() and nvhdcp_te_close() pair in
nvhdcp_dev_open(). So removing this pair when it is run as vitual machine.
It is duplicated only when using TEGRA_DC_HDCP_POLICY_ALWAYS_ON policy
in device tree.

Bug 4026730

Change-Id: Ib53e6442ca125a67148021598a8e45f464581a77
Signed-off-by: Yong Goo Yi <yyi@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-nvidia/+/2877968
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: Ahung Cheng <ahcheng@nvidia.com>
Reviewed-by: Kwangwoo Lee <kwangwool@nvidia.com>
GVS: Gerrit_Virtual_Submit <buildbot_gerritrpt@nvidia.com>